ABC_Chemicals_Model/requirements.txt
ipykernel==6.29.3
ipython==8.22.2
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
scikit-learn==1.4.1.post1
scipy==1.12.0
plotly==5.17.0
pyyaml==6.0.1
nbformat>=4.2.0
Jinja2
tabulate
